VULNERABILITY: Attempt to fix bug where slock may not come up
Service stopped attempting to start on a boot for me. Seems that it
hit a limiter, which should be disabled with =0 and avoided by default
with =3. It sounds like it should have kept trying, so maybe  a bug in
systemd, not sure. Either way, this may be a viable fix.